[ref] DRY up concept of redis instances within engine
**Problem**

Redis instances are retrieved and operated on like models, but because
they are not defined as models logic like finding is repeated.

**Solution**

Define Redis instances as a model within the engine. This encourged
a further distinction in naming with the `RedisInstance` library
concept. Renaming it to `ConfigRedisInstance` clarifies that this is
really created for configuration, not broader app consumption.

**Notes**

In the future, further decoupling between the config instances and the
model instances may be warranted, but for now it seems good enough.